APP/requirements.txt
flask
python-dotenv
requests
pytest
pytest-cov